back-forward-list: add new back-forward-list for tracking navigation.
This will be used to track all edit points and various focused widgets
within the application. We can then add keybindings to move backwards
and forwards throughout the application.

For example, when I switch to docs (with ctrl+k or something), i want to
be able to go back to my focused editor with a simple keystroke. (this is
done with ctrl+o/ctrl+i in vim).