libide: add libide design and prototype
            ** This is not ready for public consumption **

This includes the basic design and skeleton implementation of plans for
libide. Everything extends from the IdeContext object. All objects, with
the exception of IdeContext, are decendants of an IdeContext.

Since scripting is important for this system, it is important that objects
are GObject Introspectable. We will add this in the not too distant
future.

For more information on the objects and components of libide, see
libide/DESIGN.md.

If you would like to contribute to libide development, get in touch via
email or IRC so we can collaborate on which component you would like to
implement.

-- Christian