libide: add stub implementation for ide_buffer_get_line_flags()
This function will eventually track the changes in a buffer based on the
response from the Vcs. Doing so will give us a bit better abstraction
over the current git implementation.

We'll probably implement the lookup hash independently though, so it can
be resused based on diff callbacks from the VCS implementation.