libide: add snippets engine to LibIDE
This is a fairly substantial move to bring snippets into LibIDE. We need
them in the same location as IdeSourceView, and they are probably generally
useful anyway.

A follow up commit will plumb these into the IdeContext and IdeSourceView.

I've switched some of these to use the new style G_DECLARE* macros, and I'd
like to see more of that. We also need to run through the headers and audit
the includes. (Generally I only want ide-object.h/ide-types.h included in
these files if possible).

The completion provider and parsers are not public API, and therefore not
exported into the Gir.