libide: add ide_buffer_get_content()
This helper returns a GBytes containing the buffer contents. As more
subsystems need to access the buffer contents, this can help us avoid
calling get_text() so much. Especially since we need to transform the
buffer to include the trailing newline.

Furthermore, it allows us a single point to update the IdeUnsavedFiles
whenever the content is needed. This means that lost content will get
automatically persisted to the drafts when closed.