libide: be smarter about avoiding diff calculation
Try to get the things we know need to be calculated immediately served
immediately. Try to catch the hard stuff (like manually changing a line
back to its original contents) in a followup handler a second after the
most recent edit.

The longer we keep typing, the longer we try to delay the followup pass.
This is because it is pretty likely the user will perform either a newline
insertion or deletion (requiring an immediate diff) within the not too
distant future.

This significantly reduces the number of diffs calculated for me. Should
result in some power savings.