libide: setup git vcs to be reloadable
This is the first step to breaking out the git vcs into a series of
mini async operations. This helps us move forward towards being able
to reload the vcs at runtime.

One of the next major hurdles to get over is how we will apply changes to
the project tree. This simply adds the new tree to the existing tree,
which means duplicated content.

Long term, we will need to signal add/removals to the tree so that items
like a project pane can reload their content.