libide: override GtkTextView::paste_clipboard() to add line paste mode
This helps when pasting entire lines (text followed by \n). We treat the
paste as an entire line paste rather than simply pasting in position.

We might need to add some special hooks for whether or not we want this,
but in general I think it is pretty useful.

To support vim mode's P, we'll need to add a "paste-clipboard-before"
anyway, so not the end of the word if we want to split this out.