libide: add plumbing for more complex movements
This allows us to specify things such as repeat counts, and inclusive vs
exclusive movements. By pulling things into a structure, we have a bit
easier of a time adding variants without the parameters getting out of
control.

For more information on exclusvie vs inclusive, see ":help exclusive" in
Vim. Essentially, it specifies if we should move before or after the
target position.