libide: leave cursor on proper line when joining lines
This fixes the situation that you join lines and the trailing line
(which has an iter at offset zero) is not empty. We put the insertion
mark back on the same line. Bindings can use this to place the cursor
wherever they want on the line that was actually joined.