libide: stay pinned to bottom of buffer during inserts
If we are already at the bottom of the buffer, then keep the view pinned
to the bottom of the buffer. Otherwise we get into a situation where we
constantly have to keep scrolling down. Pretty handy when writing new
functions at the bottom of the document.

If someone wants to make this an option, I'm fine with that, but this
should be the default.