libide: add IdeSourceView::capture-modifier and ::clear-modifier
These signals help in the process of waiting for a modifier key. This can
be used to pass to movements that require additional input.

The capture-modifier signal will block the main loop (similar to that of
gtk_dialog_run() until the key as been pressed or focus leaves). After
which, the signal bindings can continue.