libide: track recent upstream changes in gtksourceview
We have just slight variants of how things work on these (mostly due to
where we want to place the cursor afterwards). Some of them might actually
be the same, but I don't want to make that call until I have the time to
sit down and be sure.

So for now, just override the class closure.