egg: add new EggStateMachine prototype
I'd like to be able to use this for situations like the greeter where we
need to manage potentially a bunch of little state changes. Having all of
that defined in UI would be very nice. It is also unlikely to be the
only place that is useful. I imagine dialogs like the new project dialog
and project properties could use this too. However, GtkAssistant may
be more appropriate for the create/clone/etc workflow.

This has very little project for references. So that definitely needs to
be iterated upon.

I think it shows enough promise to go forward with.