workspace: sketch out new workspace layout
This adds a new custom layout manager, GbWorkspace and GbWorkspacePane to
contain the children of the workspace.

This is not a full docking library, on purpose. It only supports animated
panes on left, right, and bottom.

It does, however, allow us to avoid using GtkPaned, which has been a
nightmare for keeping sizing information how we want.

I expect GbViewGrid will get rewritten as part of this to be a custom
multi-paned in the same way that GbWorkspace is.

I think we need to dive into GtkTextView a bit, and do some performance
work there with overzealous invalidations. For example, when shrinking
or growing the textwindow, as long as we don't render an invalid area,
we can keep using the pixel cached content. We should be more lazy about
destroying/repainting the textview. (Should help make animations smoother).

Anyway, is what it is. Very eary draft, tons to do, and will likely get
rebased often.