fuzzy: support utf8 and larger data-set sizes
We had to bump the index entry to 48-bit, which definitely hurts our
compactness a bit. But this is needed to support larger data set sizes
which I hit when testing a 1.8M entry test-case.

Also, we switch to utf8 casefolding for case insensitive indexing.
We still store relative string offsets. We could probably compress
this to character index, but it doesn't really buy us anything yet.

I also switched to using GByteArray rather than manually calling
malloc and realloc. We could probably use GStringChunk, but since
these are mostly "build and re-use" indexes, the smaller space
consumption is probably better.

We still need to add support for removing data from the index, that
way we can incrementally update the index, and just leak the removed
strings in the GByteArray.