graph: keep surface around if width/height did not change
If only the X and Y coordinates changed, we can avoid redrawing the
graph surface.

We don't actually hit this today in Builder, because the GbWorkspace
works hard to not change the width/height when animating. Also, the
X/Y stays at 0,0 relative to its GdkWindow.

But good to add nonetheless.