symbol-tree: add plumbing for symbol tree abstraction
The current symbol abstraction only allows for getting an array of
symbols. This isn't entirely useful going forward as we support more
complex languages.

This provides the interface and stub implementation for clang to
extract the symbols into a tree like interface.

You will notice that the IdeSymbolTree interface is as vague as possible.
This is meant to avoid inflating symbol objects until necessary, and
allow them to be released as early as possible.

Some impelementations may choose to hold onto a translation unit in the
IdeSymbolTree subclass and inflate from internal strings on demand.