extensions: add basic extension point api
This should feel familiar to GIOExtensionPoint but allows us to integrate
with libpeas. In particular, we want the ability to swap out extension
points when libpeas plugins are loaded or unloaded.

Consumers can listen to the IdeExtensionPoint::changed signal to be
notified when the extension point has changed the preferred implementation.
Such a situation would occur when a new plugin is loaded (or existing
plugin is unloaded) which provides an implementation at a higher priority
(indicated by a lower integer value).