extensions: use external data in libpeas plugin info
This required a lot of churn in the IdeExtensionPoint API (which is a good
thing). It also required making services and a few other base classes use
GTypeInterface instead.

However, this allows for using dependency injection from .plugin files
for a concrete implementation. We can even swap out the old version at
runtime for a higher priority version when a plugin is loaded/unloaded.