clang: extract clang into a plugin
This required some interface changes as well, which was not easy to do
as it's own commit. So this feels a bit like an amalgamation, but not
much of a cleaner way to do it.

I notice that the clang highlighter is not highlighting immediately, but
it does seem to be working (Control+L forces re-highlight) which renders
properly.