plugins: allow plugins to implement concrete extensions to buffer
This handles the ability to provide symbol resolvers and highlighters from
plugins. In particular, it handles the loading and unloading of plugins at
runtime, which should cause the implementation used to be changed.

This means that if we had two plugins say they can semantically highlight
C code, and one is unloaded, the other can take its place without having
to reload the buffer manually.